LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/iframe.http-rp/unset/worker-classic.http-expected.txt



FAIL Referrer Policy: Expects stripped-referrer for worker-classic to same-http origin and keep-origin redirection from http context. assert_equals: Reported Referrer URL is 'stripped-referrer'. expected (string) "http://localhost:8800/common/security-features/scope/document.py?policyDeliveries=%5B%5D" but got (undefined) undefined
FAIL Referrer Policy: Expects stripped-referrer for worker-classic to same-http origin and no-redirect redirection from http context. assert_equals: Reported Referrer URL is 'stripped-referrer'. expected (string) "http://localhost:8800/common/security-features/scope/document.py?policyDeliveries=%5B%5D" but got (undefined) undefined
